Fix test and address feedback
